Reorganized profile components
