+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivity/ProfileActivity.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/shared.scss';
+@import '../../../../styles/shared.scss';
 
 .ProfileActivity {
   border: 1px solid $neutral-200;
