+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivity/ProfileActivityRow/ProfileActivityRow.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/shared.scss';
+@import '../../../../../styles/shared.scss';
 
 .ProfileActivityRow {
   position: relative;
