+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivity/ProfileActivityRow/index.tsx
@@ -1,0 +1,3 @@
+import ProfileActivityRow from './ProfileActivityRow';
+
+export default ProfileActivityRow;
