+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivity/ReferralsTab/ReferralsTab.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/shared.scss';
+@import '../../../../../styles/shared.scss';
 
 .ReferralsTab {
   display: flex;
