+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileActivity/index.tsx
@@ -1,0 +1,3 @@
+import ProfileActivity from './ProfileActivity';
+
+export default ProfileActivity;
