+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileHeader/ProfileHeader.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/shared.scss';
+@import '../../../../styles/shared.scss';
 
 .ProfileHeader {
   border: 1px solid $neutral-200;
