+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileHeader/ProfileHeader.tsx
@@ -14,10 +14,10 @@ import { useInviteLinkModal } from 'state/ui/modals';
 import useUserStore from 'state/ui/user';
 import { MarkdownViewerWithFallback } from 'views/components/MarkdownViewerWithFallback/MarkdownViewerWithFallback';
 import { CWButton } from 'views/components/component_kit/new_designs/CWButton';
-import type NewProfile from '../../../models/NewProfile';
-import { SharePopover } from '../SharePopover';
-import { CWText } from '../component_kit/cw_text';
-import { SocialAccounts } from '../social_accounts';
+import type NewProfile from '../../../../models/NewProfile';
+import { SharePopover } from '../../SharePopover';
+import { CWText } from '../../component_kit/cw_text';
+import { SocialAccounts } from '../../social_accounts';
 
 type ProfileHeaderProps = {
   profile: NewProfile;
