+++ packages/commonwealth/client/scripts/views/components/Profile/ProfileHeader/index.tsx
@@ -1,0 +1,3 @@
+import ProfileHeader from './ProfileHeader';
+
+export default ProfileHeader;
